pyspark-application/requirements.txt
hdfs
elasticsearch
kafka-python